Update ADTs use in team.py
Changed the creation of ADT objects to be made in the init as them being made in add_to_team() function would override the original array each time a monster would be added.

Also created helper function __sorting_key() to find the key that should be used to sort monsters in array depending on the sort_mode the user chose. This output gets inputted into ListItem in add_to_team as the 'key' for the value being 'monster' (monster object being added to team).